[Refactor] Share the remaining duplicated oracles and search spaces
Test/benchmark oracles. tileops/testing/ already holds references imported by
both sides (mamba2_reference, gqa_fp8_utils); two more pairs join it rather
than staying verbatim copies:

- ssd_chunk_scan_fwd_ref, 60 identical lines in bench_mamba and test_mamba.
  Their docstrings had already drifted — one called prev_states dtype, the
  other float32. workloads/mamba.py builds it float32, so the test copy was
  right and the merged one keeps that.
- MHC ref_program, two pairs not one: Pre at 50 lines and Post at 10. Both
  move to tileops/testing/mhc_reference.py as free functions; the four
  methods keep their names, since TestBase.check calls ref_program, and
  delegate.

Autotune search spaces:

- Twelve byte-identical copies across gqa_fwd and gqa_bwd become
  attention/_config.tile_stage_thread_configs. Every removed line appeared
  exactly twelve times, so nothing behind the duplication differed.
- convolution's nine copies were not identical: five differ in the search
  space, one opens with a use_direct early return, one filters block_k
  against c_in. conv_autotune_configs takes the space as keyword arguments
  and does the shared-memory filtering; the two special cases keep their own
  logic at the call site rather than pushing a predicate into the helper.

Verified the generated spaces are unchanged: identical config lists for three
dtypes across all five convolution spaces, and identical ordering for the 54
GQA entries.
