+++ tileops/testing/mamba2_reference.py
@@ -93,3 +93,65 @@ def mamba2_fwd_ref(
     y_intra = torch.einsum("bchls,bchsp->bchlp", lcb, wx_t).permute(0, 1, 3, 2, 4)
 
     return (y_hist + y_intra).reshape(b, S, h, p)
+
+
+def ssd_chunk_scan_fwd_ref(x, cb, dA_cumsum, C, prev_states, dt, n_groups):
+    """Official-aligned PyTorch reference for chunk scan.
+
+    Inputs (official layouts):
+      x:           [B, S, H, P]        dtype
+      cb:          [B, C, G, L, L]     dtype    group-owned
+      dA_cumsum:   [B, H, C, L]        float32
+      C:           [B, S, G, N]        dtype    group-owned
+      prev_states: [B, C, H, P, N]     float32  P before N
+      dt:          [B, H, C, L]        dtype
+
+    Output: [B, S, H, P]  float32
+    """
+    b, S, h, p = x.shape
+    _, _, c, L = dA_cumsum.shape
+    n = C.shape[-1]
+    g = n_groups
+    heads_per_group = h // g
+
+    x_chunked = x.float().reshape(b, c, L, h, p)             # [B, C, L, H, P]
+    C_chunked = C.float().reshape(b, c, L, g, n)              # [B, C, L, G, N]
+    # broadcast C from groups to heads: [B, C, L, H, N]
+    C_heads = C_chunked[:, :, :, torch.arange(h, device=x.device) // heads_per_group, :]
+
+    # dA_cumsum: [B, H, C, L] -> [B, C, L, H] for broadcast
+    dA = dA_cumsum.float().permute(0, 2, 3, 1)  # [B, C, L, H]
+
+    # --- History path: exp(dA_l) * C[l] @ prev_states[p, n] ---
+    # prev_states: [B, C, H, P, N]
+    # C_heads:     [B, C, L, H, N] -> einsum over n: [B, C, L, H, P]
+    y_off = torch.einsum("bclhn,bchpn->bclhp", C_heads, prev_states.float())
+    y_off = y_off * torch.exp(dA).unsqueeze(-1)  # scale by exp(dA_l)
+
+    # --- Intra-chunk path: sum_{s<=l} cb[l,s] * exp(dA_l - dA_s) * dt[s] * x[s] ---
+    # cb: [B, C, G, L, L]; broadcast to heads [B, C, H, L, L]
+    cb_chunked = cb.float()  # [B, C, G, L, L]
+    cb_heads = cb_chunked[:, :, torch.arange(h, device=x.device) // heads_per_group, :, :]
+
+    # decay[b,c,h,l,s] = exp(dA_cumsum[l] - dA_cumsum[s])
+    dA_l = dA_cumsum.float().unsqueeze(-1)  # [B, H, C, L, 1]
+    dA_s = dA_cumsum.float().unsqueeze(-2)  # [B, H, C, 1, L]
+    decay = torch.exp(dA_l - dA_s)         # [B, H, C, L, L]
+
+    # causal mask
+    mask = torch.tril(torch.ones(L, L, device=x.device, dtype=torch.bool))
+    decay = decay.masked_fill(~mask.unsqueeze(0).unsqueeze(0).unsqueeze(0), 0.0)
+    decay = decay.permute(0, 2, 1, 3, 4)   # [B, C, H, L, L]
+
+    # dt: [B, H, C, L] -> [B, C, H, 1, L]
+    dt_s = dt.float().permute(0, 2, 1, 3).unsqueeze(-2)  # [B, C, H, 1, L]
+
+    # lcb[b,c,h,l,s] = cb[l,s] * decay[l,s] * dt[s]
+    lcb = cb_heads * decay * dt_s  # [B, C, H, L, L]
+
+    # y_diag[b,c,l,h,p] = sum_s lcb[b,c,h,l,s] * x[b,c,s,h,p]
+    y_diag = torch.einsum("bchls,bcshp->bclhp", lcb, x_chunked)
+
+    # combine and reshape to [B, S, H, P]
+    out = (y_off + y_diag).reshape(b, S, h, p)
+    return out
